메인페이지 컴포넌트화 작업 진행중
