+++ public/data/friendList.json
@@ -1,0 +1,17 @@
+[
+    {
+      "id": 1,
+      "userName": "hoonee",
+      "content": "회원님을 위한 추천"
+    },
+    {
+      "id": 2,
+      "userName": "su_woooooo",
+      "content": "회원님을 위한 추천"
+    },
+    {
+      "id": 3,
+      "userName": "y_jaehoon90",
+      "content": "회원님을 위한 추천"
+    }
+  ]
