+++ src/Pages/sunghoonbae/Main/Components/FeedList/comment.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class CommentInfo extends React.Component {
+    render() {
+        return(
+            <>
+              <li>{this.props.key} {this.props.name} {this.props.comment} </li>
+            </>
+        )
+    }
+}
+
+export default CommentInfo;
