+++ src/Pages/sunghoonbae/Main/Components/SideBar/friendRecommentList.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class FriendRecommentList extends React.Component {
+    render(){
+        return(
+            <>
+
+            </>
+        )
+    }
+}
+
+export default FriendRecommentList
