+++ src/Pages/sunghoonbae/Main/Main.scss
@@ -4,6 +4,7 @@
   box-sizing: border-box;
 }
 
+
 .nav {
   display: flex;
   justify-content: center;
@@ -228,7 +229,6 @@
 .flex_align_center {
   display: flex;
   align-items: center;
-  padding: 10px 10px;
 }
 
 .flex-dir {
