--- src/Pages/sunghoonbae/Main/comment.js
@@ -1,20 +1,0 @@
-import React from 'react';
-
-class CommentInfo extends React.Component {
-    render() {
-        console.log(this.props.comments)
-        return(
-            <>
-            {/* {this.props.commentInfo.map(comment => {
-                return (
-                )
-            })} */}
-                 
-              {/* {commentList} */}
-              {/* <li> {this.props.name} {this.props.comment} </li> */}
-            </>
-        )
-    }
-}
-
-export default CommentInfo;
